refactor: Reorganized files in Lib folder for better structure
Improve code maintainability and readability.
