--- app/lib/convertStringToUrlFriendly.ts
@@ -1,5 +1,0 @@
-export function stringToUrlFriendly(text: string) {
-    
-    return text.replace(/[^A-Za-z0-9]+/g, '-').toLowerCase()
-
-}
