--- app/lib/regexOnlyAlphabetic.ts
@@ -1,7 +1,0 @@
-export default function regexOnlyAlphabetic(text: string) {
-
-    const newText = text.replace(/[^a-zA-Z ]/g, " ")
-
-    return newText
-
-}
